requirements.txt
Flask
psycopg2-binary
azure-identity
azure-keyvault-secrets